requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
python-multipart
python-dotenv
pydantic-settings
alembic